final commit

+++ src/camera.cpp
@@ -1,6 +1,4 @@
 #include "camera.h"
-#include "math.h"
-#include <iostream>
 
 SF3dVector F3dVector ( GLfloat x, GLfloat y, GLfloat z ){
 	SF3dVector tmp;
@@ -26,12 +24,8 @@ void AddF3dVectorToVector ( SF3dVector * Dst, SF3dVector * V2){
 
 CCamera::CCamera(){
 	//Init with standard OGL values:
-	Position = F3dVector (	0.0, 
-							0.0,
-							0.0);
-	ViewDir = F3dVector(	0.0,
-							0.0,
-							-1.0);
+	Position = F3dVector (0.0, 0.0, 0.0);
+	ViewDir = F3dVector(0.0, 0.0, -1.0);
 	ViewDirChanged = false;
 	//Only to be sure:
 	RotatedX = RotatedY = RotatedZ = 0.0;
@@ -40,32 +34,32 @@ CCamera::CCamera(){
 void CCamera::GetViewDir( void ){
 	SF3dVector Step1, Step2;
 	//Rotate around Y-axis.
-	Step1.x = cos( (RotatedY + 90.0) * PIdiv180);
-	Step1.z = -sin( (RotatedY + 90.0) * PIdiv180);
+	Step1.x = (GLfloat) cos( (RotatedY + 90.0) * PIdiv180);
+	Step1.z = (GLfloat) -sin( (RotatedY + 90.0) * PIdiv180);
 	//Rotate around X-axis.
 	double cosX = cos (RotatedX * PIdiv180);
-	Step2.x = Step1.x * cosX;
-	Step2.z = Step1.z * cosX;
-	Step2.y = sin(RotatedX * PIdiv180);
+	Step2.x =  Step1.x * (GLfloat) cosX;
+	Step2.z =  Step1.z * (GLfloat) cosX;
+	Step2.y = (GLfloat) sin(RotatedX * PIdiv180);
 	//Rotation around Z-axis not yet implemented, so.
 	ViewDir = Step2;
 }
 
-void CCamera::Move (SF3dVector Direction){
+void CCamera::Move (SF3dVector Direction) {
 	AddF3dVectorToVector(&Position, &Direction );
 }
 
-void CCamera::RotateY (GLfloat Angle){
+void CCamera::RotateY (GLfloat Angle) {
 	RotatedY += Angle;
 	ViewDirChanged = true;
 }
 
-void CCamera::RotateX (GLfloat Angle){
+void CCamera::RotateX (GLfloat Angle) {
 	RotatedX += Angle;
 	ViewDirChanged = true;
 }
 
-void CCamera::Render( void ){
+void CCamera::Render( void ) {
 	glRotatef(-RotatedX , 1.0, 0.0, 0.0);
 	glRotatef(-RotatedY , 0.0, 1.0, 0.0);
 	glRotatef(-RotatedZ , 0.0, 0.0, 1.0);
@@ -73,12 +67,12 @@ void CCamera::Render( void ){
 }
 
 void CCamera::Orbit(GLfloat Angle, GLfloat offset, GLfloat positionX, GLfloat positionZ) {
-	Position.z = (CAMERA_RADIUS * cos((Angle)* PI / 180)) + positionZ;
-	Position.x = (CAMERA_RADIUS * sin((Angle)* PI / 180)) + positionX;
+	Position.z = (GLfloat) (CAMERA_RADIUS * cos((Angle)* PI / 180)) + positionZ;
+	Position.x = (GLfloat) (CAMERA_RADIUS * sin((Angle)* PI / 180)) + positionX;
 	RotateY(offset);
 }
 
-void CCamera::MoveForwards( GLfloat Distance ){
+void CCamera::MoveForwards( GLfloat Distance ) {
 	if (ViewDirChanged) GetViewDir();
 	SF3dVector MoveVector;
 	MoveVector.x = ViewDir.x * -Distance;
